feat: Ignore media sessions from video player apps
This commit introduces a blocklist to prevent the Dynamic Island media controls from appearing for video playback applications like YouTube, Netflix, and others.

A new `ignoredMediaPackages` set has been added to `CapsuleAccessibilityService`. When a media session update is received, the service now checks if the session's package name is in this blocklist. If it is, the media session is ignored, and the media controls will not be displayed.